Add temporary logging for T210739
Attempts to reproduce the bug locally and on mwdebug1001 have failed, so
let's add some logging in production to try to get more information.

Note the LOCK IN SHARE MODE is there to try to bypass MySQL transaction
isolation, as that's one theory as to why things are screwing up.

History:
* PS2: Determine whether it's the ID being 0 or mismatched revisions.
  Result: ID was 0.
* PS3: Check whether the page row exists on master, and whether the
  Title has the right page ID.
  Result: Page row does not exist on master, and Title has a valid ID.
* PS5: Check whether the page has archive rows, and whether/where
  the `page` deletes are being rolled back.
  Result: TBD.

Bug: T210739
Change-Id: I060d991d28ed83ae64c6be5f27cd62faaf1c710d