Clone the Title object to prevent mutations.
The Title object that is loaded from master gets reloaded from the replicas
and mutates the original object. When pages are moved, the Title no longer
exists on master, but still exists on the replicas. Cloning the object allows
the item to be loaded from the replicas, without mutating the original Title.

Bug: T210739
Change-Id: I9ad973e9a609124749909605f37bc1e1fc549585
(cherry picked from commit dedfe98eaad7015d52001988eeea9c00749e1a30)